rename newgroup.svelte
