+++ frontend/app/src/components/home/createOrUpdateGroup/GroupDetails.svelte
